Fix a out-of-range bug (that was also a mistake in the code)

+++ prog/include/objective.hpp
@@ -26,7 +26,7 @@ namespace wallin
   protected:
     Random randomVar;
     std::string name;
-    std::vector<int> heuristicValueHelper; 
+    std::vector<double> heuristicValueHelper; 
   };
 
   class NoneObj : public Objective
